backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
python-multipart
aiofiles
celery[redis]
redis
pydantic[email]
python-dotenv
psycopg2-binary
argon2-cffi
llama-index
openai
llama-parse
cohere
pandas
asyncio
ragas
deepeval
llama-index-llms-groq
llama-index-embeddings-huggingface
huggingface-hub
langchain
langchain-openai
langchain-groq
nest-asyncio
